requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
schedule>=1.2.0
flask>=3.0.0
apscheduler>=3.10.0
gunicorn>=21.2.0